+++ .buildkite/scripts/build-storybook.sh
@@ -5,5 +5,7 @@ set -e
 . ".buildkite/scripts/helpers/setup-registry.sh"
 
 yarn install --frozen-lockfile
+yarn workspace @kaizen/design-tokens prepublish
+yarn workspace @kaizen/tailwind prepublish
 yarn storybook:build
 tar -czf ./storybook.tar.gz ./storybook/public