+++ packages/components/src/KaizenProvider/ThemeProvider/themes/index.ts
@@ -1,0 +1,5 @@
+export * from "./heart"
+export * from "./types"
+
+import { heartTheme } from "./heart"
+export const defaultTheme = heartTheme