+++ packages/date-picker/src/FilterDateRangePicker/components/DateRangeInputField/DateRangeInputField.tsx
@@ -2,13 +2,17 @@ import React, { HTMLAttributes } from "react"
 import classnames from "classnames"
 import { VisuallyHidden } from "@kaizen/a11y"
 import { OverrideClassName } from "@kaizen/component-base"
-import { FieldMessage, FieldMessageStatus } from "@kaizen/draft-form"
+import { FieldMessage } from "@kaizen/draft-form"
 import { DateInput, DateInputProps } from "../../../_subcomponents/DateInput"
 import {
   DateInputDescription,
   DateInputDescriptionProps,
 } from "../../../_subcomponents/DateInputDescription"
 import { isRefObject } from "../../../utils/isRefObject"
+import {
+  DateRangeValidationMessage,
+  DateRangeValidationMessageProps,
+} from "../DateRangeValidationMessage"
 import styles from "./DateRangeInputField.module.scss"
 
 export interface DateRangeInputFieldProps
@@ -28,11 +32,11 @@ export interface DateRangeInputFieldProps
   /**
    * Updates the styling of the validation FieldMessage
    */
-  status?: FieldMessageStatus
+  status?: DateRangeValidationMessageProps["status"]
   /**
    * A descriptive message for `status` states
    */
-  validationMessage?: React.ReactNode
+  validationMessage?: DateRangeValidationMessageProps["validationMessage"]
   disabled?: boolean
 }
 
@@ -85,7 +89,7 @@ export const DateRangeInputField = React.forwardRef<
             aria-describedby={inputDescribedBy}
             autoComplete="off"
             disabled={disabled}
-            status={status}
+            status={status?.dateStart}
             {...inputRangeStartProps}
             classNameOverride={classnames(
               styles.inputRangeStart,
@@ -98,7 +102,7 @@ export const DateRangeInputField = React.forwardRef<
             aria-describedby={inputDescribedBy}
             autoComplete="off"
             disabled={disabled}
-            status={status}
+            status={status?.dateEnd}
             {...inputRangeEndProps}
             classNameOverride={classnames(
               styles.inputRangeEnd,
@@ -106,15 +110,15 @@ export const DateRangeInputField = React.forwardRef<
             )}
           />
         </fieldset>
-
-        {validationMessage && (
-          <FieldMessage
-            id={errorMessageId}
-            message={validationMessage}
+        {status && validationMessage && (
+          <DateRangeValidationMessage
             status={status}
-            reversed={isReversed}
+            validationMessage={validationMessage}
+            isReversed={isReversed}
+            id={errorMessageId}
           />
         )}
+
         <FieldMessage
           id={descriptionId}
           message={