+++ draft-packages/popover/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.28](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@5.2.27...@kaizen/draft-popover@5.2.28) (2023-02-26)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
 ## [5.2.27](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@5.2.26...@kaizen/draft-popover@5.2.27) (2023-02-23)
 
 **Note:** Version bump only for package @kaizen/draft-popover